Change `actual` to require {0,1}, other minor changes.
The original plan for `actual` to be anything that can resolve to a boolean was
causing some problems. Given that nearly every ml dataset uses {0,1} for
labels, we changed it to expect only that.

Restricted the plotting of the curves to only plot within the range of
predictions.

Also adjusted some of the variable names and plot annotations to better reflect
common ml usage.